requirements.txt
numpy<1.20
https://download.pytorch.org/whl/cu118/torch-2.1.0%2Bcu118-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cu118/torchvision-0.16.0%2Bcu118-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cu118/torchaudio-2.1.0%2Bcu118-cp39-cp39-linux_x86_64.whl


pydantic==2.4.0
pycparser==2.21.0
emoji
openai
pillow==9.5.0

transformers==4.40
ftfy 
regex 
tqdm
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/segment-anything.git

pycocotools 
matplotlib 
onnxruntime 
onnx==1.9.0
gradio==3.41.2
accelerate
bitsandbytes==0.43
easyocr
opencv-python==4.5.1.48

pytorch-lightning==2.1.2
einops
omegaconf
deepspeed
torchmetrics
webdataset
tensorboard
PyMCubes
trimesh
rembg
diffusers==0.19.3
bitsandbytes
imageio[ffmpeg]
xatlas
plyfile
xformers==0.0.22.post7
git+https://github.com/NVlabs/nvdiffrast/
huggingface-hub
langchain
langchain_community
edge-tts==6.1.12
safetensors
peft

faiss-gpu
datasets
sentencepiece
huggingface_hub
torch
spaces
